Move descriptor merging into config chain processing.
